Simple syntax errors
Code analysis (warnings) for most common dummy errors issue.  Checks
for missing semicolon before the body of a structure and for missing
parentheses in function declarations. It is not meant to be complete
(which would require a parser) but instead is designed to catch most of
this type of errors.